arreglado problema de permisos

+++ app/Http/Controllers/HomeController.php
@@ -44,16 +44,18 @@ class HomeController extends Controller
       }
 
       $tarea = Task::find($id);
-      ### evaluar el cambio de situación del estado
-      switch($estado){
-        case 1:
-          $tarea->estado = 'En proceso';
-          break;
-        case 2:
-          $tarea->estado = 'Completada';
-          break;
+      if($tarea->user_id === Auth::id()){
+        ### evaluar el cambio de situación del estado
+        switch($estado){
+          case 1:
+            $tarea->estado = 'En proceso';
+            break;
+          case 2:
+            $tarea->estado = 'Completada';
+            break;
+        }
+        $tarea->save();
       }
-      $tarea->save();
       return redirect('/home');
     }
 
@@ -63,7 +65,9 @@ class HomeController extends Controller
       }
 
       $tarea = Task::find($id);
-      $tarea->delete();
+      if($tarea->user_id === Auth::id()){
+        $tarea->delete();
+      }
       return redirect('/home');
     }
 }
